Create Double_Pendulum_GUI_Processing
This version of the GUI include an initialize variables and states button as well as a swing up button. It also includes PID displays that will show the user the current values for the Ki Kp and Kd gains. This version has errors in the display options that will be corrected in the next version